big refactor of files, simplified bundle creator
